refactor: 테스트 패키지 구조 변경
- 패키지 구조 변경에 따른 테스트 패키지 구조도 변경
