+++ backend/bottari/src/main/java/com/bottari/report/service/ReportService.java
@@ -1,14 +1,14 @@
 package com.bottari.report.service;
 
 import com.bottari.bottaritemplate.domain.BottariTemplate;
-import com.bottari.member.domain.Member;
 import com.bottari.bottaritemplate.dto.ReportBottariTemplateRequest;
+import com.bottari.bottaritemplate.repository.BottariTemplateRepository;
 import com.bottari.error.BusinessException;
 import com.bottari.error.ErrorCode;
-import com.bottari.bottaritemplate.repository.BottariTemplateRepository;
+import com.bottari.member.domain.Member;
 import com.bottari.member.repository.MemberRepository;
-import com.bottari.report.repository.ReportRepository;
 import com.bottari.report.domain.Report;
+import com.bottari.report.repository.ReportRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
