+++ backend/bottari/src/test/java/com/bottari/member/domain/MemberTest.java
@@ -1,10 +1,9 @@
-package com.bottari.member;
+package com.bottari.member.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.bottari.error.BusinessException;
-import com.bottari.member.domain.Member;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
